[module] Verplaats gesprekken naar een module
