+++ htdocs/.htaccess
@@ -38,7 +38,7 @@ RewriteRule ^forum				index.php?c=Forum [L]
 RewriteRule ^agenda				index.php?c=Agenda [L]
 RewriteRule ^courant			index.php?c=Courant [L]
 RewriteRule ^fotoalbum			index.php?c=FotoAlbum [L]
-RewriteRule ^gesprekken			index.php?c=Gesprekken [L]
+RewriteRule ^gesprekken			index.php?m=gesprekken&c=Gesprekken [L]
 #RewriteRule ^geolocation		index.php?c=GeoLocation [L]
 RewriteRule ^bijbelrooster		index.php?c=Bijbelrooster [L]
 RewriteRule ^rechten			index.php?c=Rechten [L]
